refactor(server-cloudflare): merge env types into factory.ts
Single source of truth — AppEnv, Bindings, and the factory instance
all live together. Deletes the intermediate env.ts file.
